+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-09 03:13:21
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 6daf90f0194a57e33aec460a40042fd3ad686b0d
+
+## [ForwardWidgets-2kuai] - 2025-08-09 03:13:21
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: e58e97e5953813134ca96bc7158fa52c15a39a6c
+
+## [ForwardWidgets-pack1r] - 2025-08-09 03:13:21
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 9365412fcb4707dc0c9f96d0feac6ae12257ea00
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-08 03:27:05
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main